Add NowPlayingWidget for displaying now playing movies
This commit introduces the `NowPlayingWidget` class, a stateless widget designed to display now playing movies. The widget utilizes the `BlocBuilder` to listen for state changes in the `MoviesBloc`. It conditionally renders a carousel of movies when the state is `RequestState.loaded`, otherwise it displays a skeleton loading carousel.

The widget is part of the `lib/movies/presentation/widgets` directory and is exported via the `export_manager.dart` file.