Update SVG icons and add Google Play icon
Updated SVG icons for GitHub, Google, and LinkedIn to use simpler and more modern designs. Added a new SVG icon for Google Play. Also, refactored the `HelpScreen` to `SupportScreen` and updated the `AppString` class to include a new `rating` string. Modified the `SettingsScreen` to include a new `SettingCardItem` for rating the app on the Google Play Store.